requirements.txt
altair==5.2.0
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
docopt==0.6.2
fonttools==4.46.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
importlib-metadata==6.11.0
jieba==0.42.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
numpy==1.26.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.0.76
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pip-review==1.3.0
pipreqs==0.4.13
plotly==5.18.0
prettytable==3.9.0
protobuf==4.25.1
pyarrow==14.0.1
pydeck==0.8.1b0
pyecharts==2.0.4
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
seaborn==0.13.0
selenium==3.141.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
snapshot-selenium==0.0.2
soupsieve==2.5
streamlit==1.29.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.12
wordcloud==1.9.3
yarg==0.1.9
zipp==3.17.0
